% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{MetaculR_excitement}
\alias{MetaculR_excitement}
\title{Find exciting questions}
\usage{
MetaculR_excitement(MetaculR_questions, days = 30)
}
\arguments{
\item{MetaculR_questions}{A MetaculR_questions object}

\item{days}{The time period used for the excitement calculations starts this number of days ago, prior to today. E.g., if your clock says it is day 12 and your \code{days} argument is 10, the time period is day 2 until the present.}
}
\value{
A dataframe of questions with excitement measures.
\item{id}{Question ID.}
\item{title}{Question title.}
\item{Total_Change}{Cumulative delta in time period, by probability.}
\item{Total_logodds_Change}{Cumulative delta in time period, by logodds.}
\item{Total_Change_Even}{Cumulative delta toward even odds in time period, by probability.}
\item{Total_logodds_Change_Even}{Cumulative delta toward even odds in time period, by logodds.}
}
\description{
Find exciting questions
}
\examples{
\dontrun{
questions_myPredictions_byExcitement <-
  MetaculR_excitement(
    questions_myPredictions)
}
}
