% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{MetaculR_login}
\alias{MetaculR_login}
\title{Login to Metaculus}
\usage{
MetaculR_login(api_domain = "www")
}
\arguments{
\item{api_domain}{Use "www" unless you have a custom Metaculus domain}
}
\value{
Your Metaculus_user_ID.
}
\description{
Login to Metaculus
}
\examples{
\dontrun{
Metaculus_user_id <-
  MetaculR_login()
}
}
