% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{MetaculR_markdown_table}
\alias{MetaculR_markdown_table}
\title{Easily translate R dataframes to Metaculus Markdown}
\usage{
MetaculR_markdown_table(df)
}
\arguments{
\item{df}{A dataframe.}
}
\value{
A Markdown table.
}
\description{
Easily translate R dataframes to Metaculus Markdown
}
\examples{
\dontrun{
my_data <- data.frame(Year = c(2020,2021), Value = c(6, 7.2))

MetaculR_markdown_table(my_data)
}
}
