% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{MetaculR_plot}
\alias{MetaculR_plot}
\title{Plot the history of a single question}
\usage{
MetaculR_plot(MetaculR_questions, Metacular_id, scale_binary = "prob")
}
\arguments{
\item{MetaculR_questions}{A MetaculR_questions object}

\item{Metacular_id}{The ID of the question to plot}

\item{scale_binary}{Choose "prob", "odds", or "logodds"}
}
\value{
A ggplot.
}
\description{
Plot the history of a single question
}
\examples{
\dontrun{
MetaculR_plot(
  MetaculR_questions = questions_myPredictions,
  Metacular_id = 10004)
}
}
